Helper Functions for Classification Code
This is where we perform all the steps for building our feature sets and running the machine learning classification. This code was written at an early stage in our programming careers. There are far more efficient implementations that we would use now. Most of these steps can be easily automated in scikit learn, for example.